Split up Facet.Core into Module/Term/Type modules.
