+++ src/Facet/Graph.hs
@@ -24,7 +24,8 @@ import           Data.Foldable (asum, for_)
 import qualified Data.Map as Map
 import           Data.Monoid (Endo(..))
 import qualified Data.Set as Set
-import           Facet.Core
+import           Facet.Core.Module
+import           Facet.Core.Type
 import           Facet.Name
 import           Facet.Stack
 import           Facet.Syntax
