+++ src/Facet/Notice/Elab.hs
@@ -8,7 +8,7 @@ import           Data.Semigroup (stimes)
 import qualified Facet.Carrier.Throw.Inject as L
 import qualified Facet.Carrier.Write.Inject as L
 import           Facet.Context
-import           Facet.Core (Type)
+import           Facet.Core.Type (Type)
 import           Facet.Elab as Elab
 import           Facet.Notice as Notice
 import           Facet.Pretty
